--- experiments/calc.html
@@ -1,46 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  @partial("resources/head.html")
-  <title>Calculator</title>
-
-  <!-- Local Files -->
-  <link rel="stylesheet" href="resources/scss/calc.scss">
-  <script defer src="resources/js/calc.js"></script>
-
-</head>
-<body>
-  <div class="defaultLayout">
-    @partial("resources/header.html")
-    <main>
-      <div class="alert">
-        <h2>What is it?</h2>
-        <p>This experiment is a simple calculator made with JavaScript</p>
-        <div><button onclick="closeAlert(this)">got it</button></div>
-      </div>
-      <article class="card">
-        <div class="display"><div id="displayText">0</div></div>
-
-        <button class="number" onclick="addNum(1)">1</button>
-        <button class="number" onclick="addNum(2)">2</button>
-        <button class="number" onclick="addNum(3)">3</button>
-        <button class="opera" onclick="opera('div')">&div;</button>
-
-        <button class="number" onclick="addNum(4)">4</button>
-        <button class="number" onclick="addNum(5)">5</button>
-        <button class="number" onclick="addNum(6)">6</button>
-        <button class="opera" onclick="opera('mul')">&times;</button>
-
-        <button class="number" onclick="addNum(7)">7</button>
-        <button class="number" onclick="addNum(8)">8</button>
-        <button class="number" onclick="addNum(9)">9</button>
-        <button class="opera" onclick="opera('sub')">-</button>
-
-        <button class="number" onclick="addNum(0)">0</button>
-        <button class="equal" id="equal" onclick="equal()">=</button>
-        <button class="opera" onclick="opera('sum')">+</button>
-      </article>
-    </main>
-  </div>
-</body>
-</html>
