--- experiments/index.html
@@ -1,30 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  @partial("resources/head.html")
-  <title>Experiments</title>
-
-  <!-- Local Files -->
-  <link rel="stylesheet" href="resources/scss/index.scss">
-  <script defer src="resources/js/index.js"></script>
-
-</head>
-<body>
-  <div class="defaultLayout">
-
-    @partial("resources/header.html")
-
-    <main>
-      <article class="card">
-          <h1>My<br>projects</h1>
-          <p>
-            This page stores some small projects that I created
-            to test and learn new technologies.
-            They are the experiments. <br>
-            This page is an experiment!
-          </p>
-      </article>
-    </main>
-  </div>
-</body>
-</html>
