--- experiments/resources/header.html
@@ -1,21 +1,0 @@
-<header class="defaultHeader">
-  <nav class="navContainer">
-
-    <button onclick="openSidenav()" class="sidenavTrigger">
-      <i class="material-icons">menu</i>
-    </button>
-
-    <a class="logoContainer" href="index.html">
-      <i class="material-icons">&nbsp;science&nbsp;</i>
-    </a>
-
-    <ul class="navLinks">
-      <li><a href="index.html">Home</a></li>
-      <li><a href="../index.html">Curriculum</a></li>
-      <li><a href="calc.html">Calculator</a></li>
-      <li><a href="login.html">Login</a></li>
-    </ul>
-
-    <button onclick="closeSidenav()" class="screenDarkener"></button>
-  </nav>
-</header>
