--- experiments/resources/js/index.js
@@ -1,2 +1,0 @@
-import init from './modules/_init';
-init();
