--- experiments/resources/scss/calc.scss
@@ -1,56 +1,0 @@
-@use 'modules/default.scss';
-@use 'modules/components/card';
-@use 'modules/components/alert';
-@use 'modules/components/btn' as b;
-@use 'modules/utils/variables' as v;
-@use 'modules/utils/functions' as f;
-
-.card{
-  display: grid;
-  grid-template-columns: repeat(4, 60px);
-  grid-template-rows: repeat(5, 50px);
-  gap: 10px;
-
-  background-color: white;
-}
-
-.display{
-  grid-column-start: 1;
-  grid-column-end : 5;
-
-  overflow: hidden;
-  border-bottom: solid;
-  border-color: v.$accentColor;
-  #displayText{
-    float: right;
-    font-size: f.pxToEm(30);
-    color: black;
-  }
-}
-
-@mixin btn($color:null){
-  @if $color == null {
-    @include b.btn();
-  }@else{
-    @include b.btn($color);
-  }
-  padding: 0px;
-  font-size: f.pxToEm(25);
-  color: black;
-  font-weight: 500;
-}
-
-.number{
-  @include btn(#eceff1);
-}
-
-.opera{
-  @include btn(#b0bec5);
-  font-size: f.pxToEm(30);
-}
-
-#equal{
-  @include btn;
-  font-size: f.pxToEm(40);
-  grid-column: 2 / 4;
-}
