--- experiments/resources/scss/login.scss
@@ -1,35 +1,0 @@
-@use 'modules/default.scss';
-@use 'modules/components/card';
-@use 'modules/components/alert';
-@use 'modules/components/textField';
-@use 'modules/components/btn' as b;
-@use 'modules/utils/variables' as v;
-@use 'modules/utils/functions' as f;
-
-.card{
-  display: flex;
-  align-items: center;
-  flex-direction: column;
-  gap: 10px;
-
-  background-color: white;
-  color: black;
-}
-
-#loginForm{
-  display: flex;
-  align-items: center;
-  flex-direction: column;
-  gap: 10px;
-  width: 100%;
-
-}
-
-#icon{
-  font-size: f.pxToEm(120);
-  color: v.$primaryColor;
-}
-
-#login{
-  @include b.btn;
-}
