--- experiments/resources/scss/modules/components/_btn.scss
@@ -1,32 +1,0 @@
-@use '../utils/functions' as f;
-@use '../utils/shadow' as s;
-@use '../utils/typography' as t;
-@use '../utils/variables' as v;
-
-@mixin btn($color: v.$defaultColor){
-
-  all: unset;
-  @include t.roboto;
-  @include s.shadow(4);
-
-  border-radius: 4px;
-  padding: 16px;
-
-  text-align: center;
-  font-weight:500;
-  font-size: f.pxToEm(14);
-  letter-spacing: 1.25px;
-  color: white;
-
-  text-transform: uppercase;
-  vertical-align: middle;
-
-  background-color: $color;
-  transition: background-color 300ms ease;
-  &:hover, &:focus{
-    background-color:darken($color,8%);
-  }
-  &:active{
-    background-color:darken($color,25%);
-  }
-}
