--- experiments/resources/scss/modules/components/_textField.scss
@@ -1,61 +1,0 @@
-@use '../utils/variables' as v;
-
-$textFieldAccentColor: #1976d2;
-
-.textField{
-  position: relative;
-  width: 100%;
-
-  input {
-    width: 100%;
-    height: 3rem;
-    font-size: 1rem;
-    padding-bottom: 0px;
-    padding-top: 10px;
-
-    background-color: #f5f5f5;
-    border-top-right-radius: 4px;
-    border-top-left-radius: 4px;
-
-    border: 0;
-    border-bottom: 2px solid #a2a5a7;
-    outline: none;
-    box-sizing: border-box;
-
-    transition-duration: 0.2s;
-    transition-timing-function: ease-in-out;
-    transition-property: border-bottom;
-  }
-
-  label {
-    width: 100%;
-    height: 3.5rem;
-    font-size: 1rem;
-
-    top: 0;
-    left: 0; right: 0;
-    color: #666769;
-    display: flex;
-    align-items: center;
-    position: absolute;
-    font-size: 1rem;
-    cursor: text;
-    box-sizing: border-box;
-
-    transition-duration: 0.2s;
-    transition-timing-function: ease-in-out;
-    transition-property: color, font-size, top;
-  }
-}
-
-.textField.used{
-  input{
-    border-bottom: 2px solid $textFieldAccentColor;
-  }
-  label{
-    color: $textFieldAccentColor;
-    font-size: .8rem;
-    top: -20px;
-    pointer-events: none;
-  }
-}
