--- experiments/resources/scss/modules/utils/_variables.scss
@@ -1,15 +1,0 @@
-//colors
-$primaryColor: #212121;
-$primaryColorText:white;
-$accentColor: #ff5722;
-$accentColorText: white;
-
-$defaultColor: #03a9f4;
-$defaultColorText: #6200ee;
-
-//media
-$breakMedium: 800px;
-
-//others
-$headerHeight: 64px;
-$footerHeight: 0px;
