--- experiments/resources/scss/userList.scss
@@ -1,19 +1,0 @@
-@use 'modules/default.scss';
-@use 'modules/components/card';
-@use 'modules/components/alert';
-@use 'modules/components/collection';
-@use 'modules/utils/variables' as v;
-@use 'modules/utils/functions' as f;
-
-.card{
-  display: flex;
-  align-items: center;
-  flex-direction: column;
-
-  background-color: white;
-  color: black;
-}
-
-h2{
-  margin: 0px;
-}
