--- experiments/userList.html
@@ -1,32 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  @partial("resources/head.html")
-  <title>User List</title>
-
-  <!-- Local Files -->
-  <link rel="stylesheet" href="resources/scss/userList.scss">
-  <script defer src="resources/js/userList.js"></script>
-
-</head>
-<body>
-  <div class="defaultLayout">
-
-    @partial("resources/header.html")
-
-    <main>
-      <div class="alert">
-        <h2>What is it?</h2>
-        <p>This experiment lists all accounts that have already logged in on the previous screen.</p>
-        <p>The account icon is picked up from the gravatar from the email</p>
-        <div><button onclick="closeAlert(this)">got it</button></div>
-      </div>
-      <article class="card">
-        <h2>Registered users</h2>
-        <ul class="collection" id="userList">
-        </ul>
-      </article>
-    </main>
-  </div>
-</body>
-</html>
