--- index.html
@@ -1,171 +1,0 @@
-<!DOCTYPE html>
-<html lang="pt-br">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-
-  <title>Meu Currículo</title>
-  <meta http-equiv="content-type" content="text/html" charset="UTF-8">
-  <meta name="description" content="O currículo online de Silas Henrique">
-  <meta name="author" content="Silas Henrique Alves Araújo">
-
-  <meta property="og:title" content="Meu Currículo">
-  <meta property="og:description" content="O currículo online de Silas Henrique">
-  <meta property="og:image" content="resources/images/logo.png">
-  <meta property="og:url" content="silash35.github.io">
-  <meta name="twitter:card" content="resources/images/logo.png">
-
-  <link rel="preload" href="https://fonts.googleapis.com/css2?family=Roboto&display=swap" as="style" onload="this.onload=null;this.rel='stylesheet'">
-  <noscript><link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Roboto&display=swap"></noscript>
-  <link rel="stylesheet" href="resources/style.scss">
-</head>
-<body>
-  <header>
-    <h1>Silas&nbsp;Henrique&nbsp;Alves&nbsp;Araújo</h1>
-    <hr>
-    <div> Data de nascimento: 02/02/2003</div>
-    <div>
-      <a href="mailto:silash35@gmail.com" title="E-mail de Silas Henrique">silash35@gmail.com</a>
-        - (71) 98752-9296
-        - Simões Filho/BA
-    </div>
-    <div>Objetivo: Trabalhar com tecnologia, mecânica e eletricidade</div>
-    <ul>
-      <li>Entusiasta do software livre, programação e design formado no ensino fundamental pelo Colégio Objetivo</li>
-    </ul>
-  </header>
-
-  <main>
-
-    <section>
-      <h2>Formação Acadêmica</h2>
-      <ul>
-        <li> Ensino Médio - 2018 - 4° semestre - IFBA – Campus Simões Filho</li>
-        <li>Ensino Fundamental - 2017 - Colégio Objetivo</li>
-      </ul>
-    </section>
-
-  <!--
-    </section>
-      <h2>Experiência Profissional</h2>
-      <ul>
-        <li> Nada ainda</li>
-      </ul>
-    </section>
-  -->
-
-    <section>
-      <h2>Habilidades</h2>
-      <ul>
-        <li>Edição de Vídeo</li>
-        <li>Programação em C++, Qt</li>
-        <li>
-          <a href="experiments/index.html" title="GitTrainingWall" target="_blank" rel="noreferrer">
-            FrontEnd (HTML, CSS, JS)
-          </a>
-        </li>
-        <li>Inkscape</li>
-        <li>Pacote Libre Office</li>
-        <li>
-          <a href="http://brorlandi.github.io/GitTrainingWall/" title="GitTrainingWall" target="_blank" rel="noreferrer">
-            Git e GitHub
-          </a>
-        </li>
-        <li>Linux</li>
-      </ul>
-    </section>
-
-    <section>
-      <h2>Formação Complementar</h2>
-      <ul>
-        <li>
-          <a href="https://silash35.gitlab.io/large-files/certificates/megaCursos1.pdf" title="Certificado Mega Cursos 1"target="_blank" rel="noreferrer">
-            Microsoft Windows, Microsoft Office básico,
-            Digitação, Multimídia, Corel Draw, Técnicas de Impressão,
-            Treinamento para escritório e Internet
-          </a>
-          - 56 horas
-        </li>
-        <li>
-          <a href="https://silash35.gitlab.io/large-files/certificates/megaCursos2.pdf" title="Certificado Mega Curso 2" target="_blank" rel="noreferrer">
-            Microsoft Office avançado
-          </a>
-          - 32 horas
-        </li>
-        <li>
-          <a href="https://silash35.gitlab.io/large-files/certificates/megaCursos3.pdf" title="Certificado Mega Cursos 3" target="_blank" rel="noreferrer">
-            Web Design Básico
-          </a>
-          - 48 Horas
-        </li>
-        <li>
-          <a href="https://silash35.gitlab.io/large-files/certificates/senaiLogicaDaProgramacao.pdf" title="certificado de logica da programação" target="_blank" rel="noreferrer">
-            Logica da programação
-          </a>
-          - 14 Horas
-        </li>
-        <li>
-          <a href="https://silash35.gitlab.io/large-files/certificates/senaiTIC.pdf" title="Certificado Tecnologia da Informação e Comunicação" target="_blank" rel="noreferrer">
-            Tecnologia da Informação e Comunicação
-          </a>
-          - 14 Horas
-        </li>
-        <li>
-          <a href="https://silash35.gitlab.io/large-files/certificates/senaiEmpreendedorismo.pdf" title="Certificado de Empreendedorismo" target="_blank" rel="noreferrer">
-            Empreendedorismo
-          </a>
-          - 14 Horas
-        </li>
-        <li>
-          <a href="https://silash35.gitlab.io/large-files/certificates/megaCursos4.pdf" title="Certificado Mega cursos 4 Auxiliar de administração">
-            Auxiliar de administração
-          </a>
-          - 48 Horas
-        </li>
-        <li>
-          <a href="https://silash35.gitlab.io/large-files/certificates/cursoRobotica1.pdf" title="Certificado do Curso de Robótica" target="_blank" rel="noreferrer">
-            Robótica Básica
-          </a>
-          - 32 Horas
-        </li>
-        <li>
-          <a href="https://silash35.gitlab.io/large-files/certificates/megaCursos5.pdf" title="Certificado Mega 5 Manutenção de micro">
-            Manutenção de Microcomputador
-          </a>
-          - 48 Horas
-        </li>
-      </ul>
-    </section>
-
-    <section>
-      <h2>Idiomas</h2>
-      <ul>
-        <li>
-          Português - Nativo
-        </li>
-        <li>
-          <a href="https://silash35.gitlab.io/large-files/certificates/ALLingles4.pdf" title="Currículo Ingles ALL level 4" target="_blank" rel="noreferrer">
-            Inglês - Intermediário
-          </a>
-        </li>
-      </ul>
-    </section>
-
-    <section>
-      <h2>Atividades Extracurriculares</h2>
-      <ul>
-        <li>
-          <a href="https://silash35.gitlab.io/large-files/certificates/roboticaVoluntario.pdf" title="Certificado Monitor de robótica voluntario" target="_blank" rel="noreferrer">
-            Monitor do projeto de robótica educacional no IFBA – Campus Simões Filho
-          </a>
-        </li>
-        <li>
-          Monitor do projeto ECOIFBA no IFBA – Campus Simões Filho
-        </li>
-      </ul>
-    </section>
-
-  </main>
-
-</body>
-</html>
