+++ package.json
@@ -1,20 +1,20 @@
 {
   "name": "silash35.github.io",
   "version": "1.0.0",
-  "main": "index.html",
   "repository": "https://github.com/silash35/silash35.github.io.git",
   "author": "Silas Henrique <silash35@gmail.com>",
   "license": "Unlicense",
   "private": false,
+  "main": "dist/index.html",
   "scripts": {
-    "dev": "parcel index.html",
-    "build": "parcel build index.html"
+    "dev": "parcel serve src/index.html",
+    "build": "parcel build src/index.html",
+    "postbuild": "cp -r static/* dist/"
   },
   "devDependencies": {
     "eslint": "^7.8.1",
-    "parcel-bundler": "^1.12.4",
-    "parcel-plugin-html-partials": "^0.0.6",
-    "parcel-plugin-static-files-copy": "^2.5.0",
+    "parcel": "^2.0.0-beta.1",
+    "posthtml-include": "^1.5.0",
     "sass": "^1.26.10"
   },
   "dependencies": {
