--- resources/style.scss
@@ -1,45 +1,0 @@
-@charset "utf-8";
-
-body{
-  font-family: Roboto, sans-serif;
-  font-display: swap;
-  margin: 0px;
-  min-width: 40em;
-}
-
-h1{
-  text-align: center;
-  font-size: 2em;
-}
-
-h2{
-  background-color: #3a3a3a;
-  color: #ffffff;
-  font-size: 1.5em;
-
-  text-align: center;
-}
-
-div{
-  text-align: center;
-  font-size: 1.2em;
-}
-
-ul{
-  list-style-type: square;
-  font-size: 1.2em;
-}
-
-li{
-  margin-bottom: 1.2em;
-}
-
-hr{
-  border-style: hidden;
-  display: block;
-  position: relative;
-  top: -1.1em;
-  width: 40em;
-  height: 5px;
-  background-color: #000000;
-}
