+++ src/index.html
@@ -1,0 +1,27 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <include src="resources/head.html"></include>
+
+  <title>Silas Henrique</title>
+  <link rel="stylesheet" href="resources/scss/index.scss">
+  <script defer src="resources/js/index.js"></script>
+</head>
+<body>
+  <include src="resources/header.html"></include>
+  <main>
+    <article class="presentation">
+      <section class="presentationPhoto">
+        <img src="./resources/images/profile.webp" alt="Photo of silas Henrique Alves Araújo">
+      </section>
+      <section class="presentationText">
+        <h1>Silas&nbsp;Henrique</h1>
+        <p>Free software enthusiast</p>
+        <p>Programmer</p>
+        <p>C++ Specialist</p>
+        <p>Electromechanical Student</p>
+      </section>
+    </article>
+  </main>
+</body>
+</html>
