+++ src/resources/head.html
@@ -1,0 +1,19 @@
+<meta charset="UTF-8">
+<meta name="viewport" content="width=device-width, initial-scale=1.0">
+
+<meta http-equiv="content-type" content="text/html" charset="UTF-8">
+<meta name="description" content="My personal website. Here you will find my resume, contact information and some of my projects">
+<meta name="author" content="Silas Henrique Alves Araújo">
+
+<meta property="og:title" content="Silas Henrique">
+<meta property="og:description" content="My personal website. Here you will find my resume, contact information and some of my projects">
+<meta property="og:image" content="resources/images/card.png">
+<meta property="og:url" content="silash35.github.io">
+<meta name="twitter:card" content="resources/images/card.png">
+
+<!-- CDNs -->
+<link rel="preload" href="https://fonts.googleapis.com/icon?family=Material+Icons" as="style" onload="this.onload=null;this.rel='stylesheet'">
+<noscript><link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons"></noscript>
+
+<link rel="preload" href="https://fonts.googleapis.com/css2?family=Roboto:wght@100;300;400;500;700;900&display=swap" as="style" onload="this.onload=null;this.rel='stylesheet'">
+<noscript><link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Roboto:wght@100;300;400;500;700;900&display=swap"></noscript>
