+++ src/resources/header.html
@@ -1,0 +1,19 @@
+<header class="standardHeader">
+  <nav>
+
+    <a class="logoContainer" href="index.html">
+      <img src="./resources/images/icon.webp" alt="Silas Henrique Icon">
+    </a>
+
+    <ul class="navLinks">
+      <li><a href="index.html">Home</a></li>
+      <li><a href="#">About</a></li>
+      <li><a href="#">Resume</a></li>
+      <li><a href="#">Portfolio</a></li>
+      <li><a href="#">Contact</a></li>
+    </ul>
+
+    <button onclick="closeSidenav()" class="screenDarkener"></button>
+    <button onclick="openSidenav()" class="sidenavTrigger"><i class="material-icons">menu</i></button>
+  </nav>
+</header>
