+++ src/resources/js/index.js
@@ -1,0 +1,4 @@
+import init from './modules/_init';
+init();
+
+window.init = init;
