+++ src/resources/scss/modules/_standard.scss
@@ -1,0 +1,42 @@
+@charset "utf-8";
+@use 'components/navbar';
+@use 'utils/typography' as t;
+@use 'utils/variables' as v;
+
+body{
+  @include t.roboto;
+
+  margin: 0;
+  min-height: 100vh;
+
+  background-image: url("../images/tec.webp");
+  background-color: #212121;
+  background-position: center;
+  background-repeat: no-repeat;
+  background-attachment: fixed;
+  background-size: cover;
+
+  display: grid;
+  grid: 100px auto / auto;
+  grid-template-areas: "header"
+                       "main";
+
+  @media screen and (max-width: v.$breakMedium){
+    grid: 80px auto / auto;
+    grid-template-areas: "header"
+                         "main";
+  }
+}
+
+header{
+  grid-area: header;
+}
+
+main{
+  grid-area: main;
+}
+
+img{
+  width: 100%;
+  height: auto;
+}
