+++ src/resources/scss/modules/utils/_variables.scss
@@ -1,0 +1,14 @@
+//colors
+
+/*
+$primaryColor: #212121;
+$primaryColorText:white;
+$accentColor: #ff5722;
+$accentColorText: white;
+
+$defaultColor: #03a9f4;
+$defaultColorText: #6200ee;
+*/
+
+//media
+$breakMedium: 800px;
